+++ src/test/scala/edu/ie3/simona/model/participant/ChpModelSpec.scala
@@ -50,7 +50,7 @@ class ChpModelSpec
   val chpStateNotRunning: ChpState =
     ChpState(isRunning = false, 0, Kilowatts(0), KilowattHours(0))
   val chpStateRunning: ChpState =
-    ChpState(isRunning = true, 0, Kilowatts(0), KilowattHours(0))
+    ChpState(isRunning = true, 0, Kilowatts(42), KilowattHours(42))
 
   val (storageInput, chpInput) = setupSpec()
 
@@ -145,26 +145,26 @@ class ChpModelSpec
     "Check active power after calculating next state with #chpState and heat demand #heatDemand kWh:" in {
       val testCases = Table(
         ("chpState", "storageLvl", "heatDemand", "expectedActivePower"),
-        (chpStateNotRunning, 90, 0, 0), // tests case (false, false, true)
+        (chpStateNotRunning, 70, 0, 0), // tests case (false, false, true)
         (
           chpStateNotRunning,
-          90,
-          8 * 115 + 230,
+          70,
+          8 * 115,
           95,
         ), // tests case (false, true, false)
-        (chpStateNotRunning, 90, 10, 0), // tests case (false, true, true)
-        (chpStateRunning, 90, 0, 95), // tests case (true, false, true)
-        (chpStateRunning, 90, 8 * 115, 95), // tests case (true, true, false)
-        (chpStateRunning, 90, 10, 95), // tests case (true, true, true)
+        (chpStateNotRunning, 70, 10, 0), // tests case (false, true, true)
+        (chpStateRunning, 70, 0, 95), // tests case (true, false, true)
+        (chpStateRunning, 70, 8 * 115, 95), // tests case (true, true, false)
+        (chpStateRunning, 70, 10, 95), // tests case (true, true, true)
         (
           chpStateRunning,
-          90,
+          70,
           7 * 115 + 1,
           95,
         ), // test case (_, true, false) and demand covered together with chp
         (
           chpStateRunning,
-          90,
+          70,
           9 * 115,
           95,
         ), // test case (_, true, false) and demand not covered together with chp
@@ -190,26 +190,26 @@ class ChpModelSpec
     "Check total energy after calculating next state with #chpState and heat demand #heatDemand kWh:" in {
       val testCases = Table(
         ("chpState", "storageLvl", "heatDemand", "expectedTotalEnergy"),
-        (chpStateNotRunning, 90, 0, 0), // tests case (false, false, true)
+        (chpStateNotRunning, 70, 0, 0), // tests case (false, false, true)
         (
           chpStateNotRunning,
-          90,
-          8 * 115 + 230,
+          70,
+          8 * 115,
           100,
         ), // tests case (false, true, false)
-        (chpStateNotRunning, 90, 10, 0), // tests case (false, true, true)
-        (chpStateRunning, 90, 0, 100), // tests case (true, false, true)
-        (chpStateRunning, 90, 8 * 115, 100), // tests case (true, true, false)
-        (chpStateRunning, 90, 10, 100), // tests case (true, true, true)
+        (chpStateNotRunning, 70, 10, 0), // tests case (false, true, true)
+        (chpStateRunning, 70, 0, 100), // tests case (true, false, true)
+        (chpStateRunning, 70, 8 * 115, 100), // tests case (true, true, false)
+        (chpStateRunning, 70, 10, 100), // tests case (true, true, true)
         (
           chpStateRunning,
-          90,
+          70,
           7 * 115 + 1,
           100,
         ), // test case (_, true, false) and demand covered together with chp
         (
           chpStateRunning,
-          90,
+          70,
           9 * 115,
           100,
         ), // test case (_, true, false) and demand not covered together with chp
@@ -236,35 +236,35 @@ class ChpModelSpec
     "Check storage level after calculating next state with #chpState and heat demand #heatDemand kWh:" in {
       val testCases = Table(
         ("chpState", "storageLvl", "heatDemand", "expectedStoredEnergy"),
-        (chpStateNotRunning, 90, 0, 1035), // tests case (false, false, true)
+        (chpStateNotRunning, 70, 0, 805), // tests case (false, false, true)
         (
           chpStateNotRunning,
-          90,
+          70,
           8 * 115,
-          115,
+          0,
         ), // tests case (false, true, false)
-        (chpStateNotRunning, 90, 10, 1025), // tests case (false, true, true)
-        (chpStateRunning, 90, 0, 1135), // tests case (true, false, true)
-        (chpStateRunning, 90, 8 * 115, 215), // tests case (true, true, false)
-        (chpStateRunning, 90, 10, 1125), // tests case (true, true, true)
+        (chpStateNotRunning, 70, 10, 795), // tests case (false, true, true)
+        (chpStateRunning, 70, 0, 905), // tests case (true, false, true)
+        (chpStateRunning, 70, 8 * 115, 0), // tests case (true, true, false)
+        (chpStateRunning, 70, 10, 895), // tests case (true, true, true)
         (
           chpStateRunning,
-          90,
+          70,
           806,
-          329,
+          99,
         ), // test case (_, true, false) and demand covered together with chp
         (
           chpStateRunning,
-          90,
+          70,
           9 * 115,
-          100,
+          0,
         ), // test case (_, true, false) and demand not covered together with chp
         (
           chpStateRunning,
           92,
           1,
           1150,
-        ), // test case (true, true, true) and storage volume exceeds maximum
+        ), // test case (true, true, true) CHP running, storage at lvl 92 (1058 kWh) + 100 kWh from CHP exceeds max capacity (1150 kWh).
       )
 
       forAll(testCases) {
@@ -281,48 +281,54 @@ class ChpModelSpec
     }
 
     "Check time tick and running status after calculating next state with #chpState and heat demand #heatDemand kWh:" in {
-      val testCases = Seq(
-        // (ChpState, Storage Level, Heat Demand, Expected Time Tick, Expected Running Status)
+      val testCases = Table(
+        (
+          "chpState",
+          "storageLvl",
+          "heatDemand",
+          "expectedTick",
+          "expectedRunningStatus",
+        ),
         (
           chpStateNotRunning,
-          90,
+          70,
           0,
           7200,
           false,
         ), // Test case (false, false, true)
         (
           chpStateNotRunning,
-          90,
-          8 * 115 + 230,
+          70,
+          8 * 115,
           7200,
           true,
         ), // Test case (false, true, false)
         (
           chpStateNotRunning,
-          90,
+          70,
           10,
           7200,
           false,
         ), // Test case (false, true, true)
-        (chpStateRunning, 90, 0, 7200, true), // Test case (true, false, true)
+        (chpStateRunning, 70, 0, 7200, true), // Test case (true, false, true)
         (
           chpStateRunning,
-          90,
+          70,
           8 * 115,
           7200,
           true,
         ), // Test case (true, true, false)
-        (chpStateRunning, 90, 10, 7200, true), // Test case (true, true, true)
+        (chpStateRunning, 70, 10, 7200, true), // Test case (true, true, true)
         (
           chpStateRunning,
-          90,
+          70,
           806,
           7200,
           true,
         ), // Test case (_, true, false) and demand covered together with chp
         (
           chpStateRunning,
-          90,
+          70,
           9 * 115,
           7200,
           true,
@@ -336,23 +342,22 @@ class ChpModelSpec
         ), // Test case (true, true, true) and storage volume exceeds maximum
       )
 
-      for (
+      forAll(testCases) {
         (
-          chpState,
-          storageLvl,
-          heatDemand,
-          expectedTimeTick,
-          expectedRunningStatus,
-        ) <- testCases
-      ) {
-        val chpData = buildChpRelevantData(chpState, heatDemand)
-        val thermalStorage = buildThermalStorage(storageInput, storageLvl)
-        val chpModel = buildChpModel(thermalStorage)
+            chpState,
+            storageLvl,
+            heatDemand,
+            expectedTick,
+            expectedRunningStatus,
+        ) =>
+          val chpData = buildChpRelevantData(chpState, heatDemand)
+          val thermalStorage = buildThermalStorage(storageInput, storageLvl)
+          val chpModel = buildChpModel(thermalStorage)
 
-        val nextState = chpModel.calculateNextState(chpData)
+          val nextState = chpModel.calculateNextState(chpData)
 
-        nextState.lastTimeTick shouldEqual expectedTimeTick
-        nextState.isRunning shouldEqual expectedRunningStatus
+          nextState.lastTimeTick shouldEqual expectedTick
+          nextState.isRunning shouldEqual expectedRunningStatus
       }
     }